dataapp/requirements.txt
Django>=5.2
pandas
numpy
matplotlib
seaborn
liac-arff
scikit-learn
